Provide FirebaseRemoteConfig separately in sourcesets
