+++ app/src/debug/java/org/simple/clinic/DebugClinicApp.kt
@@ -13,7 +13,6 @@ import org.simple.clinic.di.AppModule
 import org.simple.clinic.di.DaggerDebugAppComponent
 import org.simple.clinic.di.DebugAppComponent
 import org.simple.clinic.di.DebugCrashReporterModule
-import org.simple.clinic.di.DebugRemoteConfigModule
 import org.simple.clinic.main.TheActivity
 import org.simple.clinic.util.AppSignature
 import org.simple.clinic.widgets.ProxySystemKeyboardEnterToImeOption
@@ -73,7 +72,6 @@ class DebugClinicApp : ClinicApp() {
     return DaggerDebugAppComponent.builder()
         .appModule(AppModule(this))
         .crashReporterModule(DebugCrashReporterModule())
-        .remoteConfigModule(DebugRemoteConfigModule())
         .build()
   }
