--- app/src/debug/java/org/simple/clinic/di/DebugRemoteConfigModule.kt
@@ -1,20 +1,0 @@
-package org.simple.clinic.di
-
-import com.google.firebase.remoteconfig.FirebaseRemoteConfig
-import com.google.firebase.remoteconfig.FirebaseRemoteConfigSettings
-import org.simple.clinic.remoteconfig.FirebaseRemoteConfigCacheExpiration
-import org.simple.clinic.remoteconfig.RemoteConfigModule
-
-class DebugRemoteConfigModule : RemoteConfigModule() {
-  override fun remoteConfig(): FirebaseRemoteConfig {
-    return super.remoteConfig().apply {
-      // Enable developer mode so that calls are not throttled during dev.
-      // More details in FirebaseRemoteConfigCacheExpiration.
-      setConfigSettings(FirebaseRemoteConfigSettings.Builder()
-          .setDeveloperModeEnabled(true)
-          .build())
-    }
-  }
-
-  override fun remoteConfigCacheExpiration() = FirebaseRemoteConfigCacheExpiration.DEBUG
-}
